Use multiplication to produce combined quality value for languages
In rare cases, content is intended for a bilingual audience. This is why
Content-Language allows multiple languages. Deriving the qvalue from an accept
header was taking only the first language specified. However, this commit
introduces the same 'trick' as used for encoding, where the qvalues are
multiplied.

This method has the following desirable properties:

- No result of multiplication can exceed the value of 1.

- Implicit AND logic: any qvalue of 0 (unacceptable) makes a total qvalue of
  0 (unacceptable).