Add reap as dependency
On balance, the benefit of being able to provide a more useful Ring-friendly
'batteries-included' set of API functions outweighs the downsides of adding an
extra dependency. So long as the reap-assuming API is in a separate namespace,
reap can be explicity excluded by users who want to expressly use pick without
reap on the classpath.