Changed API to allow undecoded metadata in representation map
Rather than pass representations with string key values for content-type,
content-language and content-encoding, the API now allows these to be passed in
a sub-map under the ::pick/representation-metadata key.

The rationale is to allow representations to be passed as they would normally
exist, using namespaced keys to highlight the keys interesting to pick. Creating
simple representations just for the purposes of selection requires work by the
user to map the result from pick back to the actual representation.